+++ lib/glimmer/dsl/libui/shape_expression.rb
@@ -47,10 +47,10 @@ module Glimmer
         end
         
         def add_content(parent, keyword, *args, &block)
+          options = args.last.is_a?(Hash) ? args.last : {post_add_content: true}
           super
-          parent.post_add_content
+          parent&.post_add_content if options[:post_add_content]
         end
-        
       end
     end
   end