+++ lib/glimmer/libui/control_proxy.rb
@@ -126,7 +126,8 @@ module Glimmer
       ]
       
       # libui returns the contained LibUI object
-      attr_reader :parent_proxy, :libui, :args, :keyword, :block
+      attr_reader :parent_proxy, :libui, :args, :keyword, :block, :content_added
+      alias content_added? content_added
       
       def initialize(keyword, parent, args, &block)
         @keyword = keyword
@@ -338,7 +339,7 @@ module Glimmer
       alias visible= visible
       
       def content(&block)
-        Glimmer::DSL::Engine.add_content(self, Glimmer::DSL::Libui::ControlExpression.new, @keyword, &block)
+        Glimmer::DSL::Engine.add_content(self, Glimmer::DSL::Libui::ControlExpression.new, @keyword, {post_add_content: @content_added}, &block)
       end
       
       private