+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support `table` `cell_rows` explicit bidirectional data-binding (with `<=>` sign)
 - Support `table` `cell_rows` explicit bidirectional data-binding with model-based rows (not `Array`s of column cells) by expecting model attributes to match underscored column names
+- Support specifying `column_attributes` as `Array` in `table` `cell_rows` explicit bidirectional data-binding with model-based rows (not `Array`s of column cells)
 - Improve examples/basic_table_button.rb with bidirectional data-binding for `table` `cell_rows`
 - Improve examples/form_table.rb with bidirectional data-binding for `table` `cell_rows`
 - Ensure examples/meta_example.rb has examples sorted